fixed the homescreen; recent 5 timesheet entries displayed; still missing show timesheet button

+++ app/src/main/java/com/mobileexam/timesheetapp/models/LogsResponse.kt
@@ -14,7 +14,7 @@ data class LogsData(
 
 data class LogItem(
     @SerializedName("_id") val id: String,
-    @SerializedName("Total Hours 100%") val totalHours: Int,
+    @SerializedName("Total Hours 100%") val totalHours: Long,
     @SerializedName("toggle") val toggle: String,
     @SerializedName("time-in") val timeIn: Long,
     @SerializedName("time-out") val timeOut: Long,
